// types.ts
export interface Note {
  id: string;
  title: string;
  courseCode: string;
  courseName: string;
  university: string;
  professor: string;
  author: string;
  authorAvatarUrl: string;
  rating: number;
  reviews: number;
  price: number;
  previewImageUrl: string;
  description: string;
  uploadDate: string;
}